Move code for extensions of GoldilocksField to its own file.
